changed int to size_t for result row count

+++ src/common/src/postgres_connection.c
@@ -393,7 +393,7 @@ General_path_element_t* get_memory(int size, General_path_element_t *path){
 }
 
 
-General_path_element_t* noPathFound(int *count, General_path_element_t *no_path) {
+General_path_element_t* noPathFound(size_t *count, General_path_element_t *no_path) {
    count = 0;
    return NULL;
 }
