+++ src/driving_distance/src/many_to_dist_driving_distance.c
@@ -42,7 +42,7 @@ static int driving_many_to_dist_driver(
           char* sql, int64_t *start_vertex, int num,
           float8 distance,
           bool directed, bool equicost, bool has_rcost,
-          General_path_element_t **path, int *path_count) {
+          General_path_element_t **path, size_t *path_count) {
   int SPIcode;
   pgr_edge_t *edges = NULL;
   int64_t total_tuples = 0;
@@ -93,15 +93,15 @@ PGDLLEXPORT Datum
 #endif
 driving_many_to_dist(PG_FUNCTION_ARGS) {
   FuncCallContext     *funcctx;
-  int                  call_cntr;
-  int                  max_calls;
+  size_t                  call_cntr;
+  size_t                  max_calls;
   TupleDesc            tuple_desc;
   General_path_element_t  *ret_path = 0;
 
   /* stuff done only on the first call of the function */
   if (SRF_IS_FIRSTCALL()) {
       MemoryContext   oldcontext;
-      int path_count = 0;
+      size_t path_count = 0;
 
       /* create a function context for cross-call persistence */
       funcctx = SRF_FIRSTCALL_INIT();
